feat: add backward-compat re-exports and strengthen entrypoint tests
Each thin wrapper script in .agents/skills/ now re-exports the
previously-public symbols from its corresponding aims.* module and
declares __all__, following the data_quality_policy.py shim pattern.
tests/test_entrypoints.py gains three additions: backtest.py in the
help-exits-zero suite, test_script_re_exports_public_api to verify
each symbol appears in __all__ and as an attribute, and
test_script_help_via_uv_run_exits_zero to exercise the uv-run path.

Also moves all reusable logic from tools/okf_hugo_adapter.py into
src/aims/okf_hugo.py (keeping the tool as a six-line CLI wrapper) and
updates test_okf_hugo_adapter.py to import from the package module with
100% branch coverage.
