+++ .agents/skills/market-analysis/scripts/backtest.py
@@ -5,7 +5,9 @@ from __future__ import annotations
 
 import sys
 
-from aims.backtest import main
+from aims.backtest import main, run_backtest
+
+__all__ = ["main", "run_backtest"]
 
 if __name__ == "__main__":
     sys.exit(main())
