+++ .agents/skills/market-analysis/scripts/market_analysis.py
@@ -5,7 +5,21 @@ from __future__ import annotations
 
 import sys
 
-from aims.market_analysis import main
+from aims.market_analysis import (
+    OhlcvBar,
+    StooqProvider,
+    generate_artifact,
+    main,
+    score_instruments,
+)
+
+__all__ = [
+    "OhlcvBar",
+    "StooqProvider",
+    "generate_artifact",
+    "main",
+    "score_instruments",
+]
 
 if __name__ == "__main__":
     sys.exit(main())
