+++ .agents/skills/market-analysis/scripts/notify_slack.py
@@ -5,7 +5,9 @@ from __future__ import annotations
 
 import sys
 
-from aims.notifications import main
+from aims.notifications import build_success_payload, main
+
+__all__ = ["build_success_payload", "main"]
 
 if __name__ == "__main__":
     sys.exit(main())
