+++ .agents/skills/market-analysis/scripts/validate_analysis.py
@@ -3,7 +3,9 @@
 
 from __future__ import annotations
 
-from aims.validate_analysis import main
+from aims.validate_analysis import main, validate_artifact
+
+__all__ = ["main", "validate_artifact"]
 
 if __name__ == "__main__":
     main()
