+++ .agents/skills/market-analysis/scripts/validate_history.py
@@ -5,7 +5,9 @@ from __future__ import annotations
 
 import sys
 
-from aims.validate_history import main
+from aims.validate_history import main, validate
+
+__all__ = ["main", "validate"]
 
 if __name__ == "__main__":
     sys.exit(main())
