+++ .agents/skills/market-analysis/scripts/validate_instrument_mappings.py
@@ -5,7 +5,9 @@ from __future__ import annotations
 
 import sys
 
-from aims.mappings import main
+from aims.mappings import main, validate_mappings
+
+__all__ = ["main", "validate_mappings"]
 
 if __name__ == "__main__":
     sys.exit(main())
