+++ .agents/skills/update-cfd-instruments/scripts/update_cfd_instruments.py
@@ -3,7 +3,9 @@
 
 from __future__ import annotations
 
-from aims.update_cfd_instruments import main
+from aims.update_cfd_instruments import build_rows, main
+
+__all__ = ["build_rows", "main"]
 
 if __name__ == "__main__":
     main()
